Renamed from classy-hash to classy_hash.
This is based on recommendations at
http://guides.rubygems.org/name-your-gem/
