+++ lib/classy_hash.rb
